// azure-ai-speech-toolkit/src/api/ui.ts
export interface UIConfig<T> {
  /**
   * name is the identifier of the UI
   */
  name: string;
  /**
   * human readable meaningful display name of the UI
   */
  title: string;
  /**
   * placeholder in the UI
   */
  placeholder?: string;
  /**
   * prompt text providing some ask or explanation to the user
   */
  prompt?: string;
  /**
   * `step` and `totalSteps` are used to describe the progress in question flow
   * `step` is the sequence number of current question
   */
  step?: number;
  /**
   * `totalStep` is the number of questions totally
   */
  totalSteps?: number;
  /**
   * default input value
   */
  default?: T | (() => Promise<T>);

  /**
   * A function that will be called to validate input and to give a hint to the user.
   *
   * @param input The current value of the input to be validated.
   * @return A human-readable string which is presented as diagnostic message.
   * Return `undefined` when 'value' is valid.
   */
  validation?: (input: T) => string | undefined | Promise<string | undefined>;

  /**
   * Actions that can be made within the question.
   * @param An array of actions
   * @param `icon` is the icon id of the action item
   * @param `tooltip` is the hint of the action item
   * @param `command` is the command name that will be executed when current action triggered
   */
  buttons?: { icon: string; tooltip: string; command: string }[];
}

export interface ConfirmConfig extends UIConfig<boolean> {
  /**
   * display text for option true or false
   */
  transformer?: (value: boolean) => string;
}

/**
 * single selection UI config
 */
export interface SingleSelectConfig extends UIConfig<string> {
  /**
   * option array or a callback function which returns option array
   */
  options: StaticOptions | (() => Promise<StaticOptions>);

  default?: string | (() => Promise<string>);

  /**
   * This config only works for option items with `OptionItem[]` type. If `returnObject` is true, the answer value is an `OptionItem` object; otherwise, the answer value is the `id` string of the `OptionItem`.
   * In case of option items with `string[]` type, whether `returnObject` is true or false, the returned answer value is always a string.
   */
  returnObject?: boolean;

  /**
   * whether skip selection if there is only one option, default is false
   */
  skipSingleOption?: boolean;
}